+++ nutype_macros/src/common/mod.rs
@@ -1,3 +1,4 @@
+pub mod fallback;
 pub mod generate;
 pub mod models;
 pub mod parse;